+++ plugin.video.arteplussept/resources/lib/view.py
@@ -1,9 +1,14 @@
 """Manage views like home menu, dynamic menus, search, favorites..."""
 # pylint: disable=import-error
 from xbmcswift2 import xbmc
+# pylint: disable=import-error
+from xbmcswift2 import xbmcgui
 
 from . import api
+from . import hof
 from . import mapper
+from . import settings as stg
+from . import user
 
 
 def build_home_page(cached_categories, settings):
@@ -12,7 +17,7 @@ def build_home_page(cached_categories, settings):
         mapper.create_search_item(),
     ]
     try:
-        live_stream_data = api.program_video(settings.language, 'LIVE')
+        live_stream_data = api.player_video(settings.language, 'LIVE')
         live_stream_item = mapper.map_live_video(
             live_stream_data, settings.quality, '1')
         addon_menu.append(live_stream_item)
@@ -46,16 +51,16 @@ def get_cached_category(category_title, most_viewed_categories):
 
 def build_favorites(plugin, settings):
     """Build the menu for user favorites thanks to API call"""
-    return [mapper.map_artetv_video(item) for item in
-            api.get_favorites(plugin, settings.language, settings.username, settings.password) or
+    return [mapper.map_artetv_item(item) for item in api.get_favorites(
+                settings.language, user.get_cached_token(plugin, settings.username)) or
             # display an empty list in case of error. error should be display in a notification
             []]
 
 
-def add_favorite(plugin, usr, pwd, program_id, label):
+def add_favorite(plugin, usr, program_id, label):
     """Add content program_id to user favorites.
     Notify about completion success or failure with label."""
-    if 200 == api.add_favorite(plugin, usr, pwd, program_id):
+    if 200 == api.add_favorite(user.get_cached_token(plugin, usr), program_id):
         msg = plugin.addon.getLocalizedString(30025).format(label=label)
         plugin.notify(msg=msg, image='info')
     else:
@@ -63,10 +68,10 @@ def add_favorite(plugin, usr, pwd, program_id, label):
         plugin.notify(msg=msg, image='error')
 
 
-def remove_favorite(plugin, usr, pwd, program_id, label):
+def remove_favorite(plugin, usr, program_id, label):
     """Remove content program_id from user favorites.
     Notify about completion success or failure with label."""
-    if 200 == api.remove_favorite(plugin, usr, pwd, program_id):
+    if 200 == api.remove_favorite(user.get_cached_token(plugin, usr), program_id):
         msg = plugin.addon.getLocalizedString(30027).format(label=label)
         plugin.notify(msg=msg, image='info')
     else:
@@ -74,20 +79,61 @@ def remove_favorite(plugin, usr, pwd, program_id, label):
         plugin.notify(msg=msg, image='error')
 
 
+def purge_favorites(plugin, usr):
+    """Flush user favorites and notify about success or failure"""
+    purge_confirmed = xbmcgui.Dialog().yesno(
+        plugin.addon.getLocalizedString(30040),
+        plugin.addon.getLocalizedString(30043),
+        autoclose=10000)
+
+    if purge_confirmed:
+        if 200 == api.purge_favorites(user.get_cached_token(plugin, usr)):
+            plugin.notify(msg=plugin.addon.getLocalizedString(30041), image='info')
+        else:
+            plugin.notify(msg=plugin.addon.getLocalizedString(30042), image='error')
+
+
+def mark_as_watched(plugin, usr, program_id, label):
+    """
+    Get program duration and synch progress with total duration
+    in order to mark a program as watched
+    """
+    status = -1
+    try:
+        program_info = api.player_video(stg.languages[0], program_id)
+        total_time = program_info.get('attributes').get('metadata').get('duration').get('seconds')
+        status = api.sync_last_viewed(user.get_cached_token(plugin, usr), program_id, total_time)
+    # pylint: disable=broad-exception-caught
+    except Exception as excp:
+        xbmc.log(f" exception caught :{str(excp)}")
+
+    if 200 == status:
+        msg = plugin.addon.getLocalizedString(30036).format(label=label)
+        plugin.notify(msg=msg, image='info')
+    else:
+        msg = plugin.addon.getLocalizedString(30037).format(label=label)
+        plugin.notify(msg=msg, image='error')
+
+
 def build_last_viewed(plugin, settings):
     """Build the menu of user history"""
-    return [mapper.map_artetv_video(item) for item in
-            api.get_last_viewed(plugin, settings.language, settings.username, settings.password) or
+    return [mapper.map_artetv_item(item) for item in api.get_last_viewed(
+                settings.language, user.get_cached_token(plugin, settings.username)) or
             # display an empty list in case of error. error should be display in a notification
             []]
 
 
-def purge_last_viewed(plugin, usr, pwd):
+def purge_last_viewed(plugin, usr):
     """Flush user history and notify about success or failure"""
-    if 200 == api.purge_last_viewed(plugin, usr, pwd):
-        plugin.notify(msg=plugin.addon.getLocalizedString(30031), image='info')
-    else:
-        plugin.notify(msg=plugin.addon.getLocalizedString(30032), image='error')
+    purge_confirmed = xbmcgui.Dialog().yesno(
+        plugin.addon.getLocalizedString(30030),
+        plugin.addon.getLocalizedString(30033),
+        autoclose=10000)
+    if purge_confirmed:
+        if 200 == api.purge_last_viewed(user.get_cached_token(plugin, usr)):
+            plugin.notify(msg=plugin.addon.getLocalizedString(30031), image='info')
+        else:
+            plugin.notify(msg=plugin.addon.getLocalizedString(30032), image='error')
 
 
 def build_mixed_collection(kind, collection_id, settings):
@@ -110,8 +156,44 @@ def build_video_streams(program_id, settings):
         item, api.streams(kind, program_id, settings.language), settings.quality)
 
 
+def build_sibling_playlist(plugin, settings, program_id):
+    """
+    Return a pair with videos belonging to the same parent as program id
+    e.g. other episodes of a same serie, videos around the same topic
+    and the start program id of this collection i.e. program_id
+    """
+    parent_program = None
+    # get parent of prefered kind first. for the moment TV_SERIES only
+    for prefered_kind in mapper.PREFERED_KINDS:
+        # pylint: disable=cell-var-from-loop
+        parent_program = hof.find(
+            lambda parent: api.is_of_kind(parent, prefered_kind),
+            api.get_parent_collection(settings.language, program_id))
+        if parent_program:
+            break
+    # if a parent was found, then return the list of kodi playable dict.
+    if parent_program:
+        sibling_arte_items = api.collection_with_last_viewed(
+            settings.language, user.get_cached_token(plugin, settings.username, True),
+            parent_program.get('kind'), parent_program.get('programId'))
+        return mapper.map_collection_as_playlist(sibling_arte_items, program_id)
+    return None
+
+def build_collection_playlist(plugin, settings, kind, collection_id):
+    """
+    Return a pair with collection with collection_id
+    and program id of the first element in the collection
+    """
+    return mapper.map_collection_as_playlist(api.collection_with_last_viewed(
+        settings.language,
+        user.get_cached_token(plugin, settings.username, True),
+        kind, collection_id))
+
 def build_stream_url(plugin, kind, program_id, audio_slot, settings):
-    """Return URL to stream content"""
+    """
+    Return URL to stream content.
+    If the content is not available, it tries to return a related trailer or teaser.
+    """
     # first try with content
     program_stream = api.streams(kind, program_id, settings.language)
     if program_stream:
@@ -133,8 +215,7 @@ def search(plugin, settings):
     query = get_search_query(plugin)
     if not query:
         plugin.end_of_directory(succeeded=False)
-    return mapper.map_cached_categories(
-        api.search(settings.language, query))
+    return mapper.map_collection_as_menu(api.search(settings.language, query))
 
 
 def get_search_query(plugin):